Restructure to expose `ColorSchemeProvider` for consumer use
