+++ packages/theming/demo/colorSchemeProvider.stories.mdx
@@ -1,0 +1,22 @@
+import { Meta, ArgsTable, Canvas, Story, Markdown } from '@storybook/addon-docs';
+import { ColorSchemeProvider } from '@zendeskgarden/react-theming';
+import { ColorSchemeProviderStory } from './stories/ColorSchemeProviderStory';
+import README from '../README.md';
+
+<Meta title="Packages/Theming/ColorSchemeProvider" component={ColorSchemeProvider} />
+
+# API
+
+<ArgsTable />
+
+# Demo
+
+## ColorSchemeProvider
+
+<Canvas>
+  <Story name="ColorSchemeProvider" args={{ initialColorScheme: 'system' }}>
+    {args => <ColorSchemeProviderStory {...args} />}
+  </Story>
+</Canvas>
+
+<Markdown>{README}</Markdown>
