+++ packages/theming/demo/themeProvider.stories.mdx
@@ -1,6 +1,5 @@
 import { Meta, ArgsTable, Canvas, Story, Markdown } from '@storybook/addon-docs';
 import { ThemeProvider, DEFAULT_THEME, PALETTE } from '@zendeskgarden/react-theming';
-import { ThemeProviderStory } from './stories/ThemeProviderStory';
 import { PaletteStory } from './stories/PaletteStory';
 import README from '../README.md';
 
@@ -19,12 +18,8 @@ import README from '../README.md';
 ## ThemeProvider
 
 <Canvas>
-  <Story
-    name="ThemeProvider"
-    args={{ initialColorScheme: 'system' }}
-    argTypes={{ theme: { control: false } }}
-  >
-    {args => <ThemeProviderStory {...args} />}
+  <Story name="ThemeProvider" args={{ theme: DEFAULT_THEME }}>
+    {args => <ThemeProvider {...args} />}
   </Story>
 </Canvas>
